requirements.txt
numpy==1.22.4
pandas==2.0.3
Pillow==9.1.1
scikit-learn==1.3.2
scipy==1.10.1
sklearn==0.0
tqdm
torch
torchvision